Added flash arrays to Registry.EM_COMMON
ltng_fed package needs them. Or else you get a segfault
when lightning_init initializes them to zero.

Discovered this when running in distributed memory dmpar as
opposed to shared smpar. It gave me a better idea of where in
the code it was dying. Admittedly this insight came in debug
mode (configure with -d option; no optimization). debug mode was slow.